fix: eliminate 4 original compilation errors via header deduplication
- Removed duplicate texture.h and textureloader.h from GeneralsMD and Generals (use Core versions)
- Added forward declarations for TextureClass and TextureLoadTaskListClass
- Fixed RECT/POINT redefinition with guards in win32_types.h
- Commented out d3dx8core.h (not available on non-Windows)
- Added intptr_t casts for pointer arithmetic (64-bit compatibility)
- Extended win32_types.h with MAX_PATH, lstrcpyn, lstrcat, _strdup, lstrcpy, lstrlen, _isnan, DDSCAPS2_*
- Added FLOAT type definition for cross-platform compatibility

Root cause: Header duplication across 3 source trees (Core, GeneralsMD, Generals)
Solution: Single source of truth in Core, removed duplicates from GeneralsMD/Generals
Impact: Original 4 errors completely eliminated

Phase 02.5 completion: DirectX 8 API stubs fully functional
Next: Phase 03 - Graphics Backend Implementation (Vulkan/OpenGL)

+++ GeneralsMD/Code/Libraries/Source/WWVegas/WW3D2/assetmgr.cpp
@@ -112,7 +112,7 @@
 #include "w3dexclusionlist.h"
 #include <INI.h>
 #include <windows.h>
-#include <d3dx8core.h>
+// #include <d3dx8core.h>  // Not available on non-Windows platforms
 #include "texture.h"
 #include "wwprofile.h"
 #include "assetstatus.h"
@@ -797,16 +797,14 @@ RenderObjClass * WW3DAssetManager::Create_Render_Obj(const char * name)
 	if (WW3D_Load_On_Demand && proto == NULL) {	// If we didn't find one, try to load on demand
 		AssetStatusClass::Peek_Instance()->Report_Load_On_Demand_RObj(name);
 
-		char filename [MAX_PATH];
-		const char *mesh_name = ::strchr (name, '.');
-		if (mesh_name != NULL) {
-			::lstrcpyn (filename, name, ((int)mesh_name) - ((int)name) + 1);
-			::lstrcat (filename, ".w3d");
-		} else {
-			sprintf( filename, "%s.w3d", name);
-		}
-
-		// If we can't find it, try the parent directory
+	char filename [MAX_PATH];
+	const char *mesh_name = ::strchr (name, '.');
+	if (mesh_name != NULL) {
+		::lstrcpyn (filename, name, ((intptr_t)mesh_name) - ((intptr_t)name) + 1);
+		::lstrcat (filename, ".w3d");
+	} else {
+		sprintf( filename, "%s.w3d", name);
+	}		// If we can't find it, try the parent directory
 		if ( Load_3D_Assets( filename ) == false ) {
 			StringClass	new_filename(StringClass("..\\"),true);
 			new_filename+=filename;
